requirements.txt
# Core dependencies
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
wandb>=0.15.0
tqdm>=4.65.0
datasets>=2.12.0
evaluate>=0.4.0
accelerate>=0.20.0
gutenbergpy>=0.1.0
nltk>=3.8.1
spacy>=3.5.0
python-dotenv>=1.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.1.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
flake8>=6.0.0
black>=23.0.0
mypy>=1.0.0
isort>=5.12.0 